requirements.txt
amqp==2.5.2
asgiref==3.2.3
beautifulsoup4==4.8.2
billiard==3.6.3.0
CacheControl==0.12.6
cachetools==4.0.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Django==3.0.1
django-cors-headers==3.2.1
django-filters==0.2.1
django-hosts==4.0
djangorestframework==3.11.0
docopt==0.6.2
fcm-django==0.3.3
geomet==0.1.2
grpcio==1.26.0
gunicorn==20.0.4
httplib2==0.16.0
idna==2.8
importlib-metadata==1.6.0
kombu==4.6.8
msgpack==0.6.2
pep8==1.7.1
pipreqs==0.4.10
pkg-resources==0.0.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfcm==1.4.7
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2019.3
redis==3.4.1
requests==2.22.0
rsa==4.0
schedule==0.6.0
six==1.11.0
soupsieve==1.9.5
SQLAlchemy==1.3.15
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
uWSGI==2.0.18
vine==1.3.0
yarg==0.1.9
zipp==3.1.0
